@@ -2,16 +2,21 @@
   "name": "landingpage-react-template",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
+    "@auth0/auth0-react": "^1.10.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^0.27.2",
     "bootstrap": "^5.2.0",
     "emailjs-com": "^2.6.4",
-    "react": "^17.0.1",
+    "mdb-react-ui-kit": "^4.1.0",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
-    "react-dom": "^17.0.1",
-    "react-scripts": "^4.0.2",
-    "smooth-scroll": "^16.1.3"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "smooth-scroll": "^16.1.3",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
