@@ -1,17 +1,34 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import * as serviceWorker from './serviceWorker';
+import reportWebVitals from './reportWebVitals';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import Citycomponents from './oldComponents/LandingPage/LoggedIN/mainComponents/Citycomponent'
+import { Auth0Provider } from '@auth0/auth0-react';
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+    <BrowserRouter>
+      <Auth0Provider
+        domain="dev-icxzyxmy.us.auth0.com"
+        clientId="NQWKNwO9vR75Iy9yHp6eADo20Ip3wgzI"
+        redirectUri={window.location.origin}
+      >
+        <Routes>
+          <Route path='/' element={<App />}>
+            <Route path='/city' element={<Citycomponents />}>
+            </Route>
+          </Route>
+
+        </Routes >
+      </Auth0Provider>
+    </BrowserRouter>
+  </React.StrictMode >
 );
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
