@@ -1,3 +1,5 @@
+import { Outlet } from "react-router-dom"
+
 export const Features = (props) => {
   return (
     <div id='features' className='text-center'>
@@ -8,15 +10,16 @@ export const Features = (props) => {
         <div className='row'>
           {props.data
             ? props.data.map((d, i) => (
-                <div key={`${d.title}-${i}`} className='col-xs-6 col-md-3'>
-                  {' '}
-                  <i className={d.icon}></i>
-                  <h3>{d.title}</h3>
-                  <p>{d.text}</p>
-                </div>
-              ))
+              <div key={`${d.title}-${i}`} className='col-xs-6 col-md-3'>
+                {' '}
+                <i className={d.icon}></i>
+                <h3>{d.title}</h3>
+                <p>{d.text}</p>
+              </div>
+            ))
             : 'Loading...'}
         </div>
+        <Outlet />
       </div>
     </div>
   )
