@@ -1,6 +1,7 @@
-import { Navbar, NavItem } from "react-bootstrap";
+// import { Navbar, NavItem } from "react-bootstrap";
 
 import React from "react";
+import { Link } from "react-router-dom";
 export const Navigation = (props) => {
   return (
     <nav id='menu' className='navbar navbar-default navbar-fixed-top'>
@@ -19,10 +20,10 @@ export const Navigation = (props) => {
             <span className='icon-bar'></span>{' '}
           </button>
           <a className='navbar-brand page-scroll' href='#page-top'>
-          YellaJo
+            YellaJo
           </a>{' '}
         </div>
-        
+
         <div
           className='collapse navbar-collapse'
           id='bs-example-navbar-collapse-1'
@@ -46,12 +47,6 @@ export const Navigation = (props) => {
             <li>
               <a href='#portfolio' className='page-scroll'>
                 Gallery
-              </a>
-            </li>
-           
-            <li>
-              <a href='/jordan' className='page-scroll'>
-                Jordan Gallery
               </a>
             </li>
             <li>
